moves s3 image file location
related #11
