+++ test/server/data/image.js
@@ -1,7 +1,7 @@
 const test = require('tape')
 const path = require('path')
 
-const image = require('../../src/server/image.js')
+const image = require('../../../src/server/data/image.js')
 
 test('generate nameKey for s3', t => {
   const givenExt = '.txt'
